fixed warnings in PDE
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2654 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/pde/algorithm/UpDownFourOpDims.hpp
@@ -74,7 +74,7 @@ protected:
 	static const size_t maxParallelDims_ = TASKS_PARALLEL_UPDOWN;
 
 	/// Map of integer to function pointer. This is used to map the dimension situation to the relevant method handler.
-	std::map<int,MFP> fnMap;
+	std::map<size_t,MFP> fnMap;
 
 	/**
 	 * Utility method to generate the fnMap member for mappings.
